requirements.txt
streamlit
pandas
numpy
openpyxl
psycopg2-binary
supabase
python-dateutil
